Refactor navigation UI to use NavigationBar, update budget data type to Int, and improve layout handling with WindowInsets.

+++ app/src/main/java/com/dionisispx/expensetracker/domain/repository/UserPreferencesRepository.kt
@@ -3,13 +3,13 @@ package com.dionisispx.expensetracker.domain.repository
 import kotlinx.coroutines.flow.Flow
 
 interface UserPreferencesRepository {
-    val totalBudget: Flow<Float>
+    val totalBudget: Flow<Int>
     val categoryLimits: Flow<Map<String, Float>>
     val themePreference: Flow<String>
     val currencyPreference: Flow<String>
     val languagePreference: Flow<String>
 
-    suspend fun saveTotalBudget(budget: Float)
+    suspend fun saveTotalBudget(budget: Int)
     suspend fun saveCategoryLimits(limits: Map<String, Float>)
     suspend fun saveThemePreference(theme: String)
     suspend fun saveCurrencyPreference(currency: String)
